pyproject.toml
[project]
name = "alpha-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apscheduler>=3.11.0",
    "black>=25.9.0",
    "fastapi>=0.119.0",
    "flake8>=7.3.0",
    "gradio>=5.49.1",
    "httpx>=0.28.1",
    "isort>=7.0.0",
    "langchain-community>=0.4",
    "langgraph>=1.0.0",
    "nest-asyncio>=1.6.0",
    "playwright>=1.55.0",
    "pylint>=4.0.1",
    "python-dotenv>=1.1.1",
    "uvicorn[standard]>=0.38.0",
]
